Refactor: 인기 체험 컴포넌트에서 api 함수 분리
