+++ src/components/mainpage/PopularActivityList.tsx
@@ -1,31 +1,13 @@
 import { useState } from 'react';
 import { useQuery } from '@tanstack/react-query';
-import axiosInstance from '@/lib/axiosInstance';
-import { ActivityResponse } from '@/types/mainPage';
 import queryKeys from '@/api/reactQuery/queryKeys';
+import getPopularActivity from '@/api/getPopularActivity';
 import PopularActivityCard from './PopularActivityCard';
 import PopularActivityButton from './PopularActivityButton';
 import PopularCardSkeleton from '../skeletonUI/mainpage/PopularCardSkeleton';
 
-const INITIAL_VALUE = {
-  activities: [],
-  totalCount: 0,
-};
-
 const OFFSET = 3;
 
-// 인기 체험 리스트 데이터를 불러오는 함수.
-async function getPopularActivity() {
-  try {
-    const res = await axiosInstance.get<ActivityResponse>(
-      '/activities?method=offset&sort=most_reviewed&page=1&size=10',
-    );
-    return res.data;
-  } catch (e) {
-    console.error('Error: ', e);
-    return INITIAL_VALUE;
-  }
-}
 const PopularActivityList = () => {
   const [startIdx, setStartIdx] = useState(0);
